refactor and review viewModel module.
