+++ app/src/main/java/com/piconemarc/personalaccountmanager/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.material.Surface
 import androidx.compose.ui.Modifier
 import com.piconemarc.personalaccountmanager.ui.component.screen.PAMMainScreen
 import com.piconemarc.personalaccountmanager.ui.theme.PersonalAccountManagerTheme
-import com.piconemarc.viewmodel.viewModel.AppActionDispatcher
+import com.piconemarc.viewmodel.viewModel.AppViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.scopes.ActivityScoped
 
@@ -30,10 +30,11 @@ class MainActivity : ComponentActivity() {
         }
     }
 
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        val appActionDispatcher : AppActionDispatcher by viewModels()
+        val appViewModel : AppViewModel by viewModels()
 
         setContent {
             PersonalAccountManagerTheme {
@@ -41,7 +42,7 @@ class MainActivity : ComponentActivity() {
                     color = MaterialTheme.colors.secondaryVariant,
                     modifier = Modifier.fillMaxSize()
                 ) {
-                    PAMMainScreen(appActionDispatcher)
+                    PAMMainScreen(appViewModel)
                 }
             }
         }
